requirements.txt
beautifulsoup4
requests
tqdm
python-docx
PyMuPDF
nltk
python-dotenv
sentence-transformers
faiss-cpu
fastapi
uvicorn
python-telegram-bot==20.7
openai